requirements.txt
PyMuPDF
Pillow
